lex: left[] and right[] hold pointers, so they are long now
lex no longer crashes. left[] and right[] are int arrays holding two
different things -- a node index into the parse tree, and for a character
class a POINTER into the packed-character array. cfoll() does

	char *p;
	p = left[v];

and scans it for a NUL. On the VAX an int held a pointer exactly; under
LP64 it holds half of one, and lex died scanning 0x4995010. The compiler
cannot rescue this: genuine int arrays, stored into four bytes and
reloaded from four, no conversion anywhere. That is the third instance of
this exact shape, after tbl's .

The 8192-byte output was a red herring -- the last flushed stdio buffer
before the crash, not a truncation. Ruling stdio in and out (getc over
37000 bytes, fopen+fprintf+exit without fclose) was time spent on the
wrong thing; the exit status said 139 all along.

It now runs to completion and fails honestly, with malloc refusing and
lex reporting 'Too little core for state generation' plus its own
counters. PORTING.md records the two candidates -- the tables genuinely
doubled, or malloc failing at a size it should not -- and that printing
the requested size distinguishes them in one run.

+++ src/cmd/lex/ldefs.c
@@ -109,8 +109,8 @@ extern int prev;	/* previous input character */
 extern int pres;	/* present input character */
 extern int peek;	/* next input character */
 extern int *name;
-extern int *left;
-extern int *right;
+extern long *left;
+extern long *right;
 extern int *parent;
 extern char *nullstr;
 extern int tptr;
